+++ backend_test.go
@@ -41,172 +41,104 @@ func testBackendGetSetDelete(t *testing.T, backend BackendType) {
 	require.NoError(t, err)
 	defer cleanupDBDir(dirname, "testdb")
 
-	// A nonexistent key should return nil, even if the key is empty
-	item, err := db.Get([]byte(""))
-	require.NoError(t, err)
-	require.Nil(t, item)
-
-	// A nonexistent key should return nil, even if the key is nil
-	value, err := db.Get(nil)
-	require.NoError(t, err)
-	require.Nil(t, value)
-
 	// A nonexistent key should return nil.
-	key := []byte("abc")
-	value, err = db.Get(key)
+	value, err := db.Get([]byte("a"))
 	require.NoError(t, err)
 	require.Nil(t, value)
 
-	// Set empty value.
-	err = db.Set(key, []byte(""))
+	ok, err := db.Has([]byte("a"))
+	require.NoError(t, err)
+	require.False(t, ok)
+
+	// Set and get a value.
+	err = db.Set([]byte("a"), []byte{0x01})
 	require.NoError(t, err)
 
-	value, err = db.Get(key)
+	ok, err = db.Has([]byte("a"))
 	require.NoError(t, err)
-	require.NotNil(t, value)
-	require.Empty(t, value)
+	require.True(t, ok)
 
-	// Set nil value.
-	err = db.Set(key, nil)
+	value, err = db.Get([]byte("a"))
 	require.NoError(t, err)
+	require.Equal(t, []byte{0x01}, value)
 
-	value, err = db.Get(key)
-	require.NoError(t, err)
-	require.NotNil(t, value)
-	require.Empty(t, value)
-
-	// Delete.
-	err = db.Delete(key)
+	err = db.SetSync([]byte("b"), []byte{0x02})
 	require.NoError(t, err)
 
-	value, err = db.Get(key)
+	value, err = db.Get([]byte("b"))
+	require.NoError(t, err)
+	require.Equal(t, []byte{0x02}, value)
+
+	// Deleting a non-existent value is fine.
+	err = db.Delete([]byte("x"))
+	require.NoError(t, err)
+
+	err = db.DeleteSync([]byte("x"))
+	require.NoError(t, err)
+
+	// Delete a value.
+	err = db.Delete([]byte("a"))
+	require.NoError(t, err)
+
+	value, err = db.Get([]byte("a"))
 	require.NoError(t, err)
 	require.Nil(t, value)
 
-	// Delete missing key.
-	err = db.Delete([]byte{9})
+	err = db.DeleteSync([]byte("b"))
 	require.NoError(t, err)
+
+	value, err = db.Get([]byte("b"))
+	require.NoError(t, err)
+	require.Nil(t, value)
+
+	// Setting, getting, and deleting an empty key should error.
+	_, err = db.Get([]byte{})
+	require.Equal(t, errKeyEmpty, err)
+	_, err = db.Get(nil)
+	require.Equal(t, errKeyEmpty, err)
+
+	_, err = db.Has([]byte{})
+	require.Equal(t, errKeyEmpty, err)
+	_, err = db.Has(nil)
+	require.Equal(t, errKeyEmpty, err)
+
+	err = db.Set([]byte{}, []byte{0x01})
+	require.Equal(t, errKeyEmpty, err)
+	err = db.Set(nil, []byte{0x01})
+	require.Equal(t, errKeyEmpty, err)
+	err = db.SetSync([]byte{}, []byte{0x01})
+	require.Equal(t, errKeyEmpty, err)
+	err = db.SetSync(nil, []byte{0x01})
+	require.Equal(t, errKeyEmpty, err)
+
+	err = db.Delete([]byte{})
+	require.Equal(t, errKeyEmpty, err)
+	err = db.Delete(nil)
+	require.Equal(t, errKeyEmpty, err)
+	err = db.DeleteSync([]byte{})
+	require.Equal(t, errKeyEmpty, err)
+	err = db.DeleteSync(nil)
+	require.Equal(t, errKeyEmpty, err)
+
+	// Setting a nil value should error, but an empty value is fine.
+	err = db.Set([]byte("x"), nil)
+	require.Equal(t, errValueNil, err)
+	err = db.SetSync([]byte("x"), nil)
+	require.Equal(t, errValueNil, err)
+
+	err = db.Set([]byte("x"), []byte{})
+	require.NoError(t, err)
+	err = db.SetSync([]byte("x"), []byte{})
+	require.NoError(t, err)
+	value, err = db.Get([]byte("x"))
+	require.NoError(t, err)
+	require.Equal(t, []byte{}, value)
 }
 
 func TestBackendsGetSetDelete(t *testing.T) {
 	for dbType := range backends {
-		testBackendGetSetDelete(t, dbType)
-	}
-}
-
-func withDB(t *testing.T, creator dbCreator, fn func(DB)) {
-	name := fmt.Sprintf("test_%x", randStr(12))
-	dir := os.TempDir()
-	db, err := creator(name, dir)
-	require.Nil(t, err)
-	defer cleanupDBDir(dir, name)
-	fn(db)
-	db.Close()
-}
-
-func TestBackendsNilKeys(t *testing.T) {
-
-	// Test all backends.
-	for dbType, creator := range backends {
-		withDB(t, creator, func(db DB) {
-			t.Run(fmt.Sprintf("Testing %s", dbType), func(t *testing.T) {
-				expect := func(key, value []byte) {
-					if len(key) == 0 { // nil or empty
-						nilValue, err := db.Get(nil)
-						assert.NoError(t, err)
-						byteValue, err := db.Get([]byte(""))
-						assert.NoError(t, err)
-						assert.Equal(t, nilValue, byteValue)
-						hasNil, err := db.Has(nil)
-						assert.NoError(t, err)
-						hasStr, err := db.Has([]byte(""))
-						assert.NoError(t, err)
-						assert.Equal(t, hasNil, hasStr)
-					}
-					value2, err := db.Get(key)
-					assert.Equal(t, value2, value)
-					assert.NoError(t, err)
-					hasKey, err := db.Has(key)
-					assert.NoError(t, err)
-					assert.Equal(t, hasKey, value != nil)
-				}
-
-				// Not set
-				expect(nil, nil)
-
-				// Set nil value
-				err := db.Set(nil, nil)
-				require.NoError(t, err)
-				expect(nil, []byte(""))
-
-				// Set empty value
-				err = db.Set(nil, []byte(""))
-				require.NoError(t, err)
-				expect(nil, []byte(""))
-
-				// Set nil, Delete nil
-				err = db.Set(nil, []byte("abc"))
-				expect(nil, []byte("abc"))
-				require.NoError(t, err)
-				err = db.Delete(nil)
-				require.NoError(t, err)
-				expect(nil, nil)
-
-				// Set nil, Delete empty
-				err = db.Set(nil, []byte("abc"))
-				expect(nil, []byte("abc"))
-				require.NoError(t, err)
-				err = db.Delete([]byte(""))
-				require.NoError(t, err)
-				expect(nil, nil)
-
-				// Set empty, Delete nil
-				err = db.Set([]byte(""), []byte("abc"))
-				expect(nil, []byte("abc"))
-				require.NoError(t, err)
-				err = db.Delete(nil)
-				require.NoError(t, err)
-				expect(nil, nil)
-
-				// Set empty, Delete empty
-				err = db.Set([]byte(""), []byte("abc"))
-				require.NoError(t, err)
-				expect(nil, []byte("abc"))
-
-				err = db.Delete([]byte(""))
-				require.NoError(t, err)
-				expect(nil, nil)
-
-				// SetSync nil, DeleteSync nil
-				err = db.SetSync(nil, []byte("abc"))
-				require.NoError(t, err)
-				expect(nil, []byte("abc"))
-				err = db.DeleteSync(nil)
-				require.NoError(t, err)
-				expect(nil, nil)
-
-				// SetSync nil, DeleteSync empty
-				err = db.SetSync(nil, []byte("abc"))
-				require.NoError(t, err)
-				err = db.DeleteSync([]byte(""))
-				require.NoError(t, err)
-
-				// SetSync empty, DeleteSync nil
-				err = db.SetSync([]byte(""), []byte("abc"))
-				require.NoError(t, err)
-				expect(nil, []byte("abc"))
-				err = db.DeleteSync(nil)
-				require.NoError(t, err)
-				expect(nil, nil)
-
-				// SetSync empty, DeleteSync empty
-				err = db.SetSync([]byte(""), []byte("abc"))
-				require.NoError(t, err)
-				expect(nil, []byte("abc"))
-				err = db.DeleteSync([]byte(""))
-				require.NoError(t, err)
-				expect(nil, nil)
-			})
+		t.Run(string(dbType), func(t *testing.T) {
+			testBackendGetSetDelete(t, dbType)
 		})
 	}
 }
@@ -223,9 +155,8 @@ func TestGoLevelDBBackend(t *testing.T) {
 
 func TestDBIterator(t *testing.T) {
 	for dbType := range backends {
-		t.Run(fmt.Sprintf("%v", dbType), func(t *testing.T) {
+		t.Run(string(dbType), func(t *testing.T) {
 			testDBIterator(t, dbType)
-			testDBIteratorBlankKey(t, dbType)
 		})
 	}
 }
@@ -239,10 +170,20 @@ func testDBIterator(t *testing.T, backend BackendType) {
 
 	for i := 0; i < 10; i++ {
 		if i != 6 { // but skip 6.
-			err := db.Set(int642Bytes(int64(i)), nil)
+			err := db.Set(int642Bytes(int64(i)), []byte{})
 			require.NoError(t, err)
 		}
 	}
+
+	// Blank iterator keys should error
+	_, err = db.Iterator([]byte{}, nil)
+	require.Equal(t, errKeyEmpty, err)
+	_, err = db.Iterator(nil, []byte{})
+	require.Equal(t, errKeyEmpty, err)
+	_, err = db.ReverseIterator([]byte{}, nil)
+	require.Equal(t, errKeyEmpty, err)
+	_, err = db.ReverseIterator(nil, []byte{})
+	require.Equal(t, errKeyEmpty, err)
 
 	itr, err := db.Iterator(nil, nil)
 	require.NoError(t, err)
@@ -346,14 +287,9 @@ func testDBIterator(t *testing.T, backend BackendType) {
 	require.NoError(t, err)
 	verifyIterator(t, ritr, []int64{9, 8, 7, 5}, "reverse iterator to 5")
 
-	// verifyIterator(t, db.Iterator(int642Bytes(0), int642Bytes(1)), []int64{0}, "forward iterator from 0 to 1")
-
 	ritr, err = db.ReverseIterator(int642Bytes(8), int642Bytes(9))
 	require.NoError(t, err)
 	verifyIterator(t, ritr, []int64{8}, "reverse iterator from 9 (ex) to 8")
-
-	// verifyIterator(t, db.Iterator(int642Bytes(2), int642Bytes(4)), []int64{2, 3}, "forward iterator from 2 to 4")
-	// verifyIterator(t, db.Iterator(int642Bytes(4), int642Bytes(2)), []int64(nil), "forward iterator from 4 to 2")
 
 	ritr, err = db.ReverseIterator(int642Bytes(2), int642Bytes(4))
 	require.NoError(t, err)
@@ -366,72 +302,11 @@ func testDBIterator(t *testing.T, backend BackendType) {
 		[]int64(nil), "reverse iterator from 2 (ex) to 4")
 }
 
-func testDBIteratorBlankKey(t *testing.T, backend BackendType) {
-	name := fmt.Sprintf("test_%x", randStr(12))
-	dir := os.TempDir()
-	db, err := NewDB(name, backend, dir)
-	require.NoError(t, err)
-	defer cleanupDBDir(dir, name)
-
-	err = db.Set([]byte(""), []byte{0})
-	require.NoError(t, err)
-	err = db.Set([]byte("a"), []byte{1})
-	require.NoError(t, err)
-	err = db.Set([]byte("b"), []byte{2})
-	require.NoError(t, err)
-
-	value, err := db.Get([]byte(""))
-	require.NoError(t, err)
-	assert.Equal(t, []byte{0}, value)
-
-	i, err := db.Iterator(nil, nil)
-	require.NoError(t, err)
-	verifyIteratorStrings(t, i, []string{"", "a", "b"}, "forward")
-
-	i, err = db.Iterator([]byte(""), nil)
-	require.NoError(t, err)
-	verifyIteratorStrings(t, i, []string{"", "a", "b"}, "forward from blank")
-
-	i, err = db.Iterator([]byte("a"), nil)
-	require.NoError(t, err)
-	verifyIteratorStrings(t, i, []string{"a", "b"}, "forward from a")
-
-	i, err = db.Iterator([]byte(""), []byte("b"))
-	require.NoError(t, err)
-	verifyIteratorStrings(t, i, []string{"", "a"}, "forward from blank to b")
-
-	i, err = db.ReverseIterator(nil, nil)
-	require.NoError(t, err)
-	verifyIteratorStrings(t, i, []string{"b", "a", ""}, "reverse")
-
-	i, err = db.ReverseIterator([]byte(""), nil)
-	require.NoError(t, err)
-	verifyIteratorStrings(t, i, []string{"b", "a", ""}, "reverse to blank")
-
-	i, err = db.ReverseIterator([]byte(""), []byte("a"))
-	require.NoError(t, err)
-	verifyIteratorStrings(t, i, []string{""}, "reverse to blank from a")
-
-	i, err = db.ReverseIterator([]byte("a"), nil)
-	require.NoError(t, err)
-	verifyIteratorStrings(t, i, []string{"b", "a"}, "reverse to a")
-}
-
 func verifyIterator(t *testing.T, itr Iterator, expected []int64, msg string) {
 	var list []int64
 	for itr.Valid() {
 		key := itr.Key()
 		list = append(list, bytes2Int64(key))
-		itr.Next()
-	}
-	assert.Equal(t, expected, list, msg)
-}
-
-func verifyIteratorStrings(t *testing.T, itr Iterator, expected []string, msg string) {
-	var list []string
-	for itr.Valid() {
-		key := itr.Key()
-		list = append(list, string(key))
 		itr.Next()
 	}
 	assert.Equal(t, expected, list, msg)
@@ -482,23 +357,22 @@ func testDBBatch(t *testing.T, backend BackendType) {
 	require.NoError(t, batch.Close())
 	assertKeyValues(t, db, map[string][]byte{"a": {1}, "b": {2}})
 
-	// writing nil keys and values should be the same as empty keys and values
-	// FIXME CLevelDB panics here: https://github.com/jmhodges/levigo/issues/55
-	if backend != CLevelDBBackend {
-		batch = db.NewBatch()
-		err = batch.Set(nil, nil)
-		require.NoError(t, err)
-		err = batch.WriteSync()
-		require.NoError(t, err)
-		assertKeyValues(t, db, map[string][]byte{"": {}, "a": {1}, "b": {2}})
+	// empty and nil keys, as well as nil values, should be disallowed
+	batch = db.NewBatch()
+	err = batch.Set([]byte{}, []byte{0x01})
+	require.Equal(t, errKeyEmpty, err)
+	err = batch.Set(nil, []byte{0x01})
+	require.Equal(t, errKeyEmpty, err)
+	err = batch.Set([]byte("a"), nil)
+	require.Equal(t, errValueNil, err)
 
-		batch = db.NewBatch()
-		err = batch.Delete(nil)
-		require.NoError(t, err)
-		err = batch.Write()
-		require.NoError(t, err)
-		assertKeyValues(t, db, map[string][]byte{"a": {1}, "b": {2}})
-	}
+	err = batch.Delete([]byte{})
+	require.Equal(t, errKeyEmpty, err)
+	err = batch.Delete(nil)
+	require.Equal(t, errKeyEmpty, err)
+
+	err = batch.Close()
+	require.NoError(t, err)
 
 	// it should be possible to write an empty batch
 	batch = db.NewBatch()