+++ boltdb_iterator.go
@@ -13,10 +13,9 @@ import (
 type boltDBIterator struct {
 	tx *bbolt.Tx
 
-	itr           *bbolt.Cursor
-	start         []byte
-	end           []byte
-	emptyKeyValue []byte // Tracks the value of the empty key, if it exists
+	itr   *bbolt.Cursor
+	start []byte
+	end   []byte
 
 	currentKey   []byte
 	currentValue []byte
@@ -29,10 +28,6 @@ var _ Iterator = (*boltDBIterator)(nil)
 
 // newBoltDBIterator creates a new boltDBIterator.
 func newBoltDBIterator(tx *bbolt.Tx, start, end []byte, isReverse bool) *boltDBIterator {
-	// We can check for empty key at the start, because we use a read/write transaction that blocks
-	// the entire database for writes while the iterator exists. If we change to a read-only txn
-	// that supports concurrency we'll need to rewrite this logic.
-	emptyKeyValue := tx.Bucket(bucket).Get(boltDBEmptyKey)
 	itr := tx.Bucket(bucket).Cursor()
 
 	var ck, cv []byte
@@ -40,31 +35,13 @@ func newBoltDBIterator(tx *bbolt.Tx, start, end []byte, isReverse bool) *boltDBI
 		switch {
 		case end == nil:
 			ck, cv = itr.Last()
-		case len(end) == 0:
-			// If end is the blank key, then we don't return any keys by definition
-			ck = nil
-			cv = nil
 		default:
 			_, _ = itr.Seek(end) // after key
 			ck, cv = itr.Prev()  // return to end key
 		}
-		// If we're currently positioned at the placeholder for the empty key, skip it (handle later)
-		if emptyKeyValue != nil && bytes.Equal(ck, boltDBEmptyKey) {
-			ck, cv = itr.Prev()
-		}
-		// If we didn't find any initial key, but there's a placeholder for the empty key at the
-		// end that we've skipped, then the initial key should be the empty one (the final one).
-		if emptyKeyValue != nil && ck == nil && (end == nil || len(end) > 0) {
-			ck = []byte{}
-			cv = emptyKeyValue
-			emptyKeyValue = nil // ensure call to Next() skips this
-		}
 	} else {
 		switch {
-		case (start == nil || len(start) == 0) && emptyKeyValue != nil:
-			ck = []byte{}
-			cv = emptyKeyValue
-		case (start == nil || len(start) == 0) && emptyKeyValue == nil:
+		case start == nil:
 			ck, cv = itr.First()
 		default:
 			ck, cv = itr.Seek(start)
@@ -72,15 +49,14 @@ func newBoltDBIterator(tx *bbolt.Tx, start, end []byte, isReverse bool) *boltDBI
 	}
 
 	return &boltDBIterator{
-		tx:            tx,
-		itr:           itr,
-		start:         start,
-		end:           end,
-		emptyKeyValue: emptyKeyValue,
-		currentKey:    ck,
-		currentValue:  cv,
-		isReverse:     isReverse,
-		isInvalid:     false,
+		tx:           tx,
+		itr:          itr,
+		start:        start,
+		end:          end,
+		currentKey:   ck,
+		currentValue: cv,
+		isReverse:    isReverse,
+		isInvalid:    false,
 	}
 }
 
@@ -127,24 +103,8 @@ func (itr *boltDBIterator) Next() {
 	itr.assertIsValid()
 	if itr.isReverse {
 		itr.currentKey, itr.currentValue = itr.itr.Prev()
-		if itr.emptyKeyValue != nil && itr.currentKey == nil {
-			// If we reached the end, but there exists an empty key whose placeholder we skipped,
-			// we should set up the empty key and its value as the final pair.
-			itr.currentKey = []byte{}
-			itr.currentValue = itr.emptyKeyValue
-			itr.emptyKeyValue = nil // This ensures the next call to Next() terminates
-		}
 	} else {
-		if len(itr.currentKey) == 0 {
-			// If the first key was the empty key, then we need to move to the first non-empty key
-			itr.currentKey, itr.currentValue = itr.itr.First()
-		} else {
-			itr.currentKey, itr.currentValue = itr.itr.Next()
-		}
-	}
-	// If we encounter the placeholder for the empty key, skip it
-	if itr.emptyKeyValue != nil && bytes.Equal(itr.currentKey, boltDBEmptyKey) {
-		itr.Next()
+		itr.currentKey, itr.currentValue = itr.itr.Next()
 	}
 }