+++ memdb.go
@@ -34,12 +34,12 @@ func (i *item) Less(other btree.Item) bool {
 
 // newKey creates a new key item.
 func newKey(key []byte) *item {
-	return &item{key: nonNilBytes(key)}
+	return &item{key: key}
 }
 
 // newPair creates a new pair item.
 func newPair(key, value []byte) *item {
-	return &item{key: nonNilBytes(key), value: nonNilBytes(value)}
+	return &item{key: key, value: value}
 }
 
 // MemDB is an in-memory database backend using a B-tree for storage.
@@ -65,6 +65,9 @@ func NewMemDB() *MemDB {
 
 // Get implements DB.
 func (db *MemDB) Get(key []byte) ([]byte, error) {
+	if len(key) == 0 {
+		return nil, errKeyEmpty
+	}
 	db.mtx.RLock()
 	defer db.mtx.RUnlock()
 
@@ -77,6 +80,9 @@ func (db *MemDB) Get(key []byte) ([]byte, error) {
 
 // Has implements DB.
 func (db *MemDB) Has(key []byte) (bool, error) {
+	if len(key) == 0 {
+		return false, errKeyEmpty
+	}
 	db.mtx.RLock()
 	defer db.mtx.RUnlock()
 
@@ -85,6 +91,12 @@ func (db *MemDB) Has(key []byte) (bool, error) {
 
 // Set implements DB.
 func (db *MemDB) Set(key []byte, value []byte) error {
+	if len(key) == 0 {
+		return errKeyEmpty
+	}
+	if value == nil {
+		return errValueNil
+	}
 	db.mtx.Lock()
 	defer db.mtx.Unlock()
 
@@ -104,6 +116,9 @@ func (db *MemDB) SetSync(key []byte, value []byte) error {
 
 // Delete implements DB.
 func (db *MemDB) Delete(key []byte) error {
+	if len(key) == 0 {
+		return errKeyEmpty
+	}
 	db.mtx.Lock()
 	defer db.mtx.Unlock()
 
@@ -161,11 +176,17 @@ func (db *MemDB) NewBatch() Batch {
 // Iterator implements DB.
 // Takes out a read-lock on the database until the iterator is closed.
 func (db *MemDB) Iterator(start, end []byte) (Iterator, error) {
+	if (start != nil && len(start) == 0) || (end != nil && len(end) == 0) {
+		return nil, errKeyEmpty
+	}
 	return newMemDBIterator(db, start, end, false), nil
 }
 
 // ReverseIterator implements DB.
 // Takes out a read-lock on the database until the iterator is closed.
 func (db *MemDB) ReverseIterator(start, end []byte) (Iterator, error) {
+	if (start != nil && len(start) == 0) || (end != nil && len(end) == 0) {
+		return nil, errKeyEmpty
+	}
 	return newMemDBIterator(db, start, end, true), nil
 }