+++ memdb_batch.go
@@ -34,6 +34,12 @@ func newMemDBBatch(db *MemDB) *memDBBatch {
 
 // Set implements Batch.
 func (b *memDBBatch) Set(key, value []byte) error {
+	if len(key) == 0 {
+		return errKeyEmpty
+	}
+	if value == nil {
+		return errValueNil
+	}
 	if b.ops == nil {
 		return errBatchClosed
 	}
@@ -43,6 +49,9 @@ func (b *memDBBatch) Set(key, value []byte) error {
 
 // Delete implements Batch.
 func (b *memDBBatch) Delete(key []byte) error {
+	if len(key) == 0 {
+		return errKeyEmpty
+	}
 	if b.ops == nil {
 		return errBatchClosed
 	}