+++ prefixdb.go
@@ -24,6 +24,9 @@ func NewPrefixDB(db DB, prefix []byte) *PrefixDB {
 
 // Get implements DB.
 func (pdb *PrefixDB) Get(key []byte) ([]byte, error) {
+	if len(key) == 0 {
+		return nil, errKeyEmpty
+	}
 	pdb.mtx.Lock()
 	defer pdb.mtx.Unlock()
 
@@ -37,6 +40,9 @@ func (pdb *PrefixDB) Get(key []byte) ([]byte, error) {
 
 // Has implements DB.
 func (pdb *PrefixDB) Has(key []byte) (bool, error) {
+	if len(key) == 0 {
+		return false, errKeyEmpty
+	}
 	pdb.mtx.Lock()
 	defer pdb.mtx.Unlock()
 
@@ -50,6 +56,12 @@ func (pdb *PrefixDB) Has(key []byte) (bool, error) {
 
 // Set implements DB.
 func (pdb *PrefixDB) Set(key []byte, value []byte) error {
+	if len(key) == 0 {
+		return errKeyEmpty
+	}
+	if value == nil {
+		return errValueNil
+	}
 	pdb.mtx.Lock()
 	defer pdb.mtx.Unlock()
 
@@ -62,6 +74,12 @@ func (pdb *PrefixDB) Set(key []byte, value []byte) error {
 
 // SetSync implements DB.
 func (pdb *PrefixDB) SetSync(key []byte, value []byte) error {
+	if len(key) == 0 {
+		return errKeyEmpty
+	}
+	if value == nil {
+		return errValueNil
+	}
 	pdb.mtx.Lock()
 	defer pdb.mtx.Unlock()
 
@@ -70,6 +88,9 @@ func (pdb *PrefixDB) SetSync(key []byte, value []byte) error {
 
 // Delete implements DB.
 func (pdb *PrefixDB) Delete(key []byte) error {
+	if len(key) == 0 {
+		return errKeyEmpty
+	}
 	pdb.mtx.Lock()
 	defer pdb.mtx.Unlock()
 
@@ -78,6 +99,9 @@ func (pdb *PrefixDB) Delete(key []byte) error {
 
 // DeleteSync implements DB.
 func (pdb *PrefixDB) DeleteSync(key []byte) error {
+	if len(key) == 0 {
+		return errKeyEmpty
+	}
 	pdb.mtx.Lock()
 	defer pdb.mtx.Unlock()
 
@@ -86,6 +110,9 @@ func (pdb *PrefixDB) DeleteSync(key []byte) error {
 
 // Iterator implements DB.
 func (pdb *PrefixDB) Iterator(start, end []byte) (Iterator, error) {
+	if (start != nil && len(start) == 0) || (end != nil && len(end) == 0) {
+		return nil, errKeyEmpty
+	}
 	pdb.mtx.Lock()
 	defer pdb.mtx.Unlock()
 
@@ -106,6 +133,9 @@ func (pdb *PrefixDB) Iterator(start, end []byte) (Iterator, error) {
 
 // ReverseIterator implements DB.
 func (pdb *PrefixDB) ReverseIterator(start, end []byte) (Iterator, error) {
+	if (start != nil && len(start) == 0) || (end != nil && len(end) == 0) {
+		return nil, errKeyEmpty
+	}
 	pdb.mtx.Lock()
 	defer pdb.mtx.Unlock()