+++ prefixdb_batch.go
@@ -16,12 +16,21 @@ func newPrefixBatch(prefix []byte, source Batch) prefixDBBatch {
 
 // Set implements Batch.
 func (pb prefixDBBatch) Set(key, value []byte) error {
+	if len(key) == 0 {
+		return errKeyEmpty
+	}
+	if value == nil {
+		return errValueNil
+	}
 	pkey := append(cp(pb.prefix), key...)
 	return pb.source.Set(pkey, value)
 }
 
 // Delete implements Batch.
 func (pb prefixDBBatch) Delete(key []byte) error {
+	if len(key) == 0 {
+		return errKeyEmpty
+	}
 	pkey := append(cp(pb.prefix), key...)
 	return pb.source.Delete(pkey)
 }