+++ prefixdb_test.go
@@ -6,28 +6,25 @@ import (
 	"github.com/stretchr/testify/require"
 )
 
-//nolint:errcheck
-func mockDBWithStuff() DB {
+func mockDBWithStuff(t *testing.T) DB {
 	db := NewMemDB()
 	// Under "key" prefix
-	db.Set(bz("key"), bz("value"))
-	db.Set(bz("key1"), bz("value1"))
-	db.Set(bz("key2"), bz("value2"))
-	db.Set(bz("key3"), bz("value3"))
-	db.Set(bz("something"), bz("else"))
-	db.Set(bz(""), bz(""))
-	db.Set(bz("k"), bz("val"))
-	db.Set(bz("ke"), bz("valu"))
-	db.Set(bz("kee"), bz("valuu"))
+	require.NoError(t, db.Set(bz("key"), bz("value")))
+	require.NoError(t, db.Set(bz("key1"), bz("value1")))
+	require.NoError(t, db.Set(bz("key2"), bz("value2")))
+	require.NoError(t, db.Set(bz("key3"), bz("value3")))
+	require.NoError(t, db.Set(bz("something"), bz("else")))
+	require.NoError(t, db.Set(bz("k"), bz("val")))
+	require.NoError(t, db.Set(bz("ke"), bz("valu")))
+	require.NoError(t, db.Set(bz("kee"), bz("valuu")))
 	return db
 }
 
 func TestPrefixDBSimple(t *testing.T) {
-	db := mockDBWithStuff()
+	db := mockDBWithStuff(t)
 	pdb := NewPrefixDB(db, bz("key"))
 
 	checkValue(t, pdb, bz("key"), nil)
-	checkValue(t, pdb, bz(""), bz("value"))
 	checkValue(t, pdb, bz("key1"), nil)
 	checkValue(t, pdb, bz("1"), bz("value1"))
 	checkValue(t, pdb, bz("key2"), nil)
@@ -41,67 +38,24 @@ func TestPrefixDBSimple(t *testing.T) {
 }
 
 func TestPrefixDBIterator1(t *testing.T) {
-	db := mockDBWithStuff()
+	db := mockDBWithStuff(t)
 	pdb := NewPrefixDB(db, bz("key"))
 
 	itr, err := pdb.Iterator(nil, nil)
 	require.NoError(t, err)
 	checkDomain(t, itr, nil, nil)
-	checkItem(t, itr, bz(""), bz("value"))
-	checkNext(t, itr, true)
 	checkItem(t, itr, bz("1"), bz("value1"))
 	checkNext(t, itr, true)
 	checkItem(t, itr, bz("2"), bz("value2"))
 	checkNext(t, itr, true)
 	checkItem(t, itr, bz("3"), bz("value3"))
 	checkNext(t, itr, false)
-	checkInvalid(t, itr)
-	itr.Close()
-}
-
-func TestPrefixDBIterator2(t *testing.T) {
-	db := mockDBWithStuff()
-	pdb := NewPrefixDB(db, bz("key"))
-
-	itr, err := pdb.Iterator(nil, bz(""))
-	require.NoError(t, err)
-	checkDomain(t, itr, nil, bz(""))
-	checkInvalid(t, itr)
-	itr.Close()
-}
-
-func TestPrefixDBIterator3(t *testing.T) {
-	db := mockDBWithStuff()
-	pdb := NewPrefixDB(db, bz("key"))
-
-	itr, err := pdb.Iterator(bz(""), nil)
-	require.NoError(t, err)
-	checkDomain(t, itr, bz(""), nil)
-	checkItem(t, itr, bz(""), bz("value"))
-	checkNext(t, itr, true)
-	checkItem(t, itr, bz("1"), bz("value1"))
-	checkNext(t, itr, true)
-	checkItem(t, itr, bz("2"), bz("value2"))
-	checkNext(t, itr, true)
-	checkItem(t, itr, bz("3"), bz("value3"))
-	checkNext(t, itr, false)
-	checkInvalid(t, itr)
-	itr.Close()
-}
-
-func TestPrefixDBIterator4(t *testing.T) {
-	db := mockDBWithStuff()
-	pdb := NewPrefixDB(db, bz("key"))
-
-	itr, err := pdb.Iterator(bz(""), bz(""))
-	require.NoError(t, err)
-	checkDomain(t, itr, bz(""), bz(""))
 	checkInvalid(t, itr)
 	itr.Close()
 }
 
 func TestPrefixDBReverseIterator1(t *testing.T) {
-	db := mockDBWithStuff()
+	db := mockDBWithStuff(t)
 	pdb := NewPrefixDB(db, bz("key"))
 
 	itr, err := pdb.ReverseIterator(nil, nil)
@@ -112,56 +66,13 @@ func TestPrefixDBReverseIterator1(t *testing.T) {
 	checkItem(t, itr, bz("2"), bz("value2"))
 	checkNext(t, itr, true)
 	checkItem(t, itr, bz("1"), bz("value1"))
-	checkNext(t, itr, true)
-	checkItem(t, itr, bz(""), bz("value"))
 	checkNext(t, itr, false)
-	checkInvalid(t, itr)
-	itr.Close()
-}
-
-func TestPrefixDBReverseIterator2(t *testing.T) {
-	db := mockDBWithStuff()
-	pdb := NewPrefixDB(db, bz("key"))
-
-	itr, err := pdb.ReverseIterator(bz(""), nil)
-	require.NoError(t, err)
-	checkDomain(t, itr, bz(""), nil)
-	checkItem(t, itr, bz("3"), bz("value3"))
-	checkNext(t, itr, true)
-	checkItem(t, itr, bz("2"), bz("value2"))
-	checkNext(t, itr, true)
-	checkItem(t, itr, bz("1"), bz("value1"))
-	checkNext(t, itr, true)
-	checkItem(t, itr, bz(""), bz("value"))
-	checkNext(t, itr, false)
-	checkInvalid(t, itr)
-	itr.Close()
-}
-
-func TestPrefixDBReverseIterator3(t *testing.T) {
-	db := mockDBWithStuff()
-	pdb := NewPrefixDB(db, bz("key"))
-
-	itr, err := pdb.ReverseIterator(nil, bz(""))
-	require.NoError(t, err)
-	checkDomain(t, itr, nil, bz(""))
-	checkInvalid(t, itr)
-	itr.Close()
-}
-
-func TestPrefixDBReverseIterator4(t *testing.T) {
-	db := mockDBWithStuff()
-	pdb := NewPrefixDB(db, bz("key"))
-
-	itr, err := pdb.ReverseIterator(bz(""), bz(""))
-	require.NoError(t, err)
-	checkDomain(t, itr, bz(""), bz(""))
 	checkInvalid(t, itr)
 	itr.Close()
 }
 
 func TestPrefixDBReverseIterator5(t *testing.T) {
-	db := mockDBWithStuff()
+	db := mockDBWithStuff(t)
 	pdb := NewPrefixDB(db, bz("key"))
 
 	itr, err := pdb.ReverseIterator(bz("1"), nil)
@@ -178,7 +89,7 @@ func TestPrefixDBReverseIterator5(t *testing.T) {
 }
 
 func TestPrefixDBReverseIterator6(t *testing.T) {
-	db := mockDBWithStuff()
+	db := mockDBWithStuff(t)
 	pdb := NewPrefixDB(db, bz("key"))
 
 	itr, err := pdb.ReverseIterator(bz("2"), nil)
@@ -193,15 +104,13 @@ func TestPrefixDBReverseIterator6(t *testing.T) {
 }
 
 func TestPrefixDBReverseIterator7(t *testing.T) {
-	db := mockDBWithStuff()
+	db := mockDBWithStuff(t)
 	pdb := NewPrefixDB(db, bz("key"))
 
 	itr, err := pdb.ReverseIterator(nil, bz("2"))
 	require.NoError(t, err)
 	checkDomain(t, itr, nil, bz("2"))
 	checkItem(t, itr, bz("1"), bz("value1"))
-	checkNext(t, itr, true)
-	checkItem(t, itr, bz(""), bz("value"))
 	checkNext(t, itr, false)
 	checkInvalid(t, itr)
 	itr.Close()